pyproject.toml
[tool.poetry]
name = "vortoj-servilo"
version = "0.1.0"
description = "Backend for the Vortoj game"
authors = ["Baptiste Darthenay <baptiste@darthenay.fr>"]
license = "Proprietary"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.8"
django = "^4"
django-anymail = {extras = ["sendgrid"], version = "^8.5"}
django-cors-headers = "^3.11.0"
django-extensions = "^3.1.5"
django-filter = "^21.1"
djangorestframework = "^3.13"
psycopg2-binary = "^2.9"
python-decouple = "^3.6"

[tool.poetry.dev-dependencies]
black = "22.3.0"
bpython = "*"
django-stubs = "^1.9.0"
djhtml = "^1.5.0"
Faker = "^13.3.3"
flake8 = "^4.0.1"
isort = "^5.10.1"
model-bakery = "^1.4.0"
pre-commit = "^2.17.0"
pytest = "^6"
pytest-cov = "^3.0.0"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.1.0"
freezegun = "^1.2.1"
